+++ test/unit/plugins/pre_commit/checks/migration_test.rb
@@ -22,17 +22,17 @@ describe PreCommit::Checks::Migration do
 
   it "succeeds if there is a migration and a schema change" do
     in_new_directory do
-      write "db/migrate/01_foo.rb", "Yep"
-      write "db/schema.rb", "version 01 bla"
-      check.call(['db/migrate/01_foo.rb', 'db/schema.rb']).must_equal nil
+      write "db/migrate/20140718171920_foo.rb", "Yep"
+      write "db/schema.rb", "version 20140718171920 bla"
+      check.call(['db/migrate/20140718171920_foo.rb', 'db/schema.rb']).must_equal nil
     end
   end
 
   it "succeeds if there is a migration and a sql schema change" do
     in_new_directory do
-      write "db/migrate/01_foo.rb", "Yep"
-      write "db/foo_structure.sql", "version 01 bla"
-      check.call(['db/migrate/01_foo.rb', 'db/foo_structure.sql']).must_equal nil
+      write "db/migrate/20140718171920_foo.rb", "Yep"
+      write "db/foo_structure.sql", "version 20140718171920 bla"
+      check.call(['db/migrate/20140718171920_foo.rb', 'db/foo_structure.sql']).must_equal nil
     end
   end
 
@@ -40,20 +40,30 @@ describe PreCommit::Checks::Migration do
     check.call(['public/javascript/foo.js', 'lib/bar.rb']).must_equal nil
   end
 
+  it "succeeds when initial schema is added with version 0" do
+    in_new_directory do
+      write "db/schema.rb", "brand new 0 version"
+      check.call(['db/schema.rb']).must_equal nil
+    end
+  end
+
   it "fails if schema change is missing" do
-    check.call(['db/migrate/01_foo.rb']).must_equal "It looks like you're adding a migration, but did not update the schema file"
+    check.call(['db/migrate/20140718171920_foo.rb']).must_equal "It looks like you're adding a migration, but did not update the schema file"
   end
 
   it "fails if migration is missing" do
-    check.call(['db/schema.rb']).must_equal "You're trying to change the schema without adding a migration file"
+    in_new_directory do
+      write "db/schema.rb", "Nope 20130111131344"
+      check.call(['db/schema.rb']).must_equal "You're trying to change the schema without adding a migration file"
+    end
   end
 
   it "fails if the schema change does not include the added versions" do
     in_new_directory do
-      write "db/migrate/123_foo.rb", "Nope"
-      write "db/migrate/234_foo.rb", "Nope"
-      write "db/schema.rb", "Nope"
-      check.call(['db/schema.rb', 'db/migrate/234_foo.rb', 'db/migrate/123_foo.rb']).must_equal "You did not add the schema versions for 234, 123 to db/schema.rb"
+      write "db/migrate/20140718171920_foo.rb", "Nope"
+      write "db/migrate/20140819201057_foo.rb", "Nope"
+      write "db/schema.rb", "Nope 20130111131344"
+      check.call(['db/schema.rb', 'db/migrate/20140819201057_foo.rb', 'db/migrate/20140718171920_foo.rb']).must_equal "You did not add the schema versions for 20140819201057, 20140718171920 to db/schema.rb"
     end
   end
 end